+++ test/markdown.spec.js
@@ -97,6 +97,10 @@ describe( 'markdown', () => {
             [ '3.', '3.' ],
             [ '3. ', '3. ' ],
             [ '3. a', '3. a' ],
+            // correct superscript and subscript
+            [ 'm<sup>2</sup>' ],
+            [ 'H<sub>2</sub>O' ],
+            [ 'H<sub onclick="alert(\"gotcha!\")">2</sub>O', 'H<sub>2</sub>O' ],
             // correct combos
             [ 'format __s__ and _e_\nformat **s** and *e*',
                 'format <strong>s</strong> and <em>e</em><br>format <strong>s</strong> and <em>e</em>'